Updated version number for the first build
Explanation for the version system:
-----------------------------------
This is the versioning system I will use in all mods from now on. The three last numbers in version number are the number of the beta build (only released on GitHub). They only change in the main java class and build.gradle, not in mcmod.info (it only shows the main version number, not the build number).